pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[dependency-groups]
dev = [
  {include-group = "test"},
  "ipython>=9.5.0",
  "mypy>=1.17.1",
  "pre-commit>=4.3.0",
  "ruff>=0.12.11",
  "types-networkx>=3.5.0.20250918",
  "types-pyyaml>=6.0.12.20250822",
  "aiida-pythonjob>=0.4.5"
]
test = [
  "pytest>=8.4.1",
  "aiida-firecrest",
  "devtools>=0.12.2"
]

[project]
authors = [
  {name = "Rico Häuselmann", email = "rico.haeuselmann@gmail.com"}
]
dependencies = [
  "aiida-core>=2.7.1",
  "cattrs[pyyaml]>=25.2.0",
  "click>=8.1.0",
  "networkx>=3.5",
  "platformdirs>=4.4.0",
  "pyyaml>=6.0.2",
  "textual",
  "tomlkit>=0.13.3",
  "typer>=0.17.3",
  "uv>=0.9.8"
]
description = "A labbook for all who run exploratory simulation experiments on HPC clusters."
name = "hpc-labbook"
readme = "README.md"
requires-python = ">=3.12"
version = "0.1.0"

[project.entry-points."aiida.calculations"]
"hpclb.generic" = "hpclb.aiida.calcjob:GenericCalculation"

[project.entry-points."aiida.parsers"]
"hpclb.generic" = "hpclb.aiida.calcjob:GenericParser"

[project.entry-points."aiida.workchains"]
"hpclb.async" = "hpclb.aiida.future:AsyncWorkchain"

[project.scripts]
hpclb = "hpclb.cli:app"

[tool.hatch.build]
packages = ["src/hpclb"]

[tool.pyright]
executionEnvironments = [{root = "src"}]
venv = "types"
venvPath = ".venv"

[tool.pytest.ini_options]
addopts = ["--doctest-modules", "--strict-markers"]
filterwarnings = [
  "ignore:Object of type .* not in session, .* operation along .* will not proceed:sqlalchemy.exc.SAWarning",
  "ignore:There is no current event loop:DeprecationWarning:plumpy",
  "ignore:Use 'content=<...>' to upload raw bytes/text content.:DeprecationWarning:httpx",
  "ignore:Creating AiiDA configuration folder.*:UserWarning"
]
markers = [
  "cluster"  # requires the local firecrestv2 docker compose cluster to be running
]

[tool.ruff]
target-version = "py312"

[tool.ruff.format]
docstring-code-format = true
quote-style = "double"

[tool.ruff.lint]
ignore = ["D203", "D212"]
select = ["E", "F", "B", "ERA", "ANN", "S", "BLE", "C4", "ICN", "PT", "Q", "RSE", "RET", "SIM", "TID", "TD", "ARG", "PTH", "FLY", "I", "C90", "N", "PERF", "D", "RUF", "TRY", "FURB"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]

[tool.uv.sources]
aiida-firecrest = {git = "https://github.com/aiidateam/aiida-firecrest.git"}
